# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Car rental API",
    "description": "\nA RESTful Car Rental API that allows users to register, authenticate, and rent cars\nfrom different locations. The system supports role-based access control with\nadmin and user roles. Admins can manage users and rental locations, while\nauthenticated users can browse locations and perform rental-related operations.\nAuthentication is handled using JWT Bearer tokens.\n     ",
    "version": "1.0.0"
  },
  "host": "localhost:4400",
  "schemes": ["http"],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "authorization",
      "in": "header",
      "description": "Authorization needed, Enter: Bearer <token>"
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/cars": {
      "get": {
        "description": "Get all cars",
        "security": [],
        "responses": { "200": { "description": "Success" } }
      },
      "post": {
        "description": "Create a new car (admin only)",
        "parameters": [{ "name": "body", "in": "body", "required": true, "schema": { "$ref": "#/definitions/Car" } }],
        "responses": { "201": { "description": "Car created successfully" } }
      }
    },
    "/api/cars/{id}": {
      "get": {
        "description": "Get car by ID",
        "parameters": [{ "name": "id", "in": "path", "required": true, "type": "string" }],
        "responses": { "200": { "description": "Success" } }
      }
    }
  },
  "definitions": {
...
    "Car": {
      "type": "object",
      "required": ["make", "model", "year", "licensePlate", "color", "seats", "mileage", "location", "pricePerDay"],
      "properties": {
        "make": { "type": "string", "example": "Toyota" },
        "model": { "type": "string", "example": "Camry" },
        "year": { "type": "integer", "example": 2023 },
        "licensePlate": { "type": "string", "example": "ABC-123" },
        "pricePerDay": { "type": "number", "example": 49.99 }
      }
    }
  }
}